alpha-arena-backend/requirements.txt
ccxt>=4.5.0
python-binance>=1.0.19
pandas>=2.3.0
numpy>=2.0.0
ta>=0.11.0
openai>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.0
rich>=13.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=8.0.0
pytest-cov>=4.0.0